+++ src/io/TiffUtils.hpp
@@ -191,7 +191,7 @@ namespace TiffUtils {
 
             // read current directory
             for (tstrip_t strip = 0; strip < TIFFNumberOfStrips(aTiff.iFile); ++strip) {
-                tmsize_t readLen = TIFFReadEncodedStrip(aTiff.iFile, strip, (&aInputMesh.mesh[0] + currentOffset), (tsize_t) -1 /* read as much as possible */);
+                tmsize_t readLen = TIFFReadEncodedStrip(aTiff.iFile, strip, (&aInputMesh.mesh[(size_t)0] + currentOffset), (tsize_t) -1 /* read as much as possible */);
                 currentOffset += readLen/sizeof(T);
             }
         }
@@ -250,7 +250,7 @@ namespace TiffUtils {
 
             size_t dataLen = ScanlineSize * height; // length of single image
             for (tstrip_t strip = 0; strip < TIFFNumberOfStrips(tif); ++strip) {
-                tmsize_t writeLen = TIFFWriteEncodedStrip(tif, strip, (void *) (&aData.mesh[0] + currentOffset), dataLen >= StripSize ? StripSize : dataLen);
+                tmsize_t writeLen = TIFFWriteEncodedStrip(tif, strip, (void *) (&aData.mesh[(size_t)0] + currentOffset), dataLen >= StripSize ? StripSize : dataLen);
                 dataLen -= writeLen;
                 currentOffset += writeLen/sizeof(T);
             }
